update tests and fix bugs
