+++ src/gbmtsplits/__init__.py
@@ -1,4 +1,5 @@
-#from .split import RandomGloballyBalancedSplit, DissimilarityDrivenGloballyBalancedSplit, ScaffoldDrivenGloballyBalancedSplit
-from .splits import GloballyBalancedSplit
+from .split import GloballyBalancedSplit
 from .clustering import *
-__version__ = '0.1.0'
+
+
+__version__ = '0.2.0'
